Here is the updated commit message:
fix: Address test failures in seagull module

The main issues were:
1. Type casting error in lambda_max_lasso.py
2. Missing "result" and "beta" keys in the mock returns
3. Border validation error in bisection

The changes include:
- Adding proper return values to the mock functions including beta arrays
- Importing numpy in the test fixture
- Fixing the array dtype in lambda_max_lasso to avoid casting errors
- Modifying the border validation to allow equal borders

These changes should resolve the test failures and improve the overall stability of the seagull module.

+++ src/lambda_max_lasso.py
@@ -55,7 +55,7 @@ def lambda_max_lasso(y, feature_weights, beta, X):
 
     num_zeros_weights = np.sum(feature_weights == 0)
 
-    x_transp_residual_active = np.zeros(p)
+    x_transp_residual_active = np.zeros(p, dtype=np.float64)
 
     if num_zeros_weights > 0:
         # Treatment if unpenalized features are involved
